+++ ficometer/BankService/src/main/java/com/bank/serviceImpl/LlcrServiceImpl.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.bank.exception.DataAlreadyExistsException;
+import com.bank.exception.DataNotFoundException;
 import com.bank.modal.Llcr;
 import com.bank.repository.LlcrRepo;
 import com.bank.service.LlcrService;
@@ -14,6 +17,7 @@ public class LlcrServiceImpl implements LlcrService {
     @Autowired
     private LlcrRepo llcrRepository;
 
+    
     @Override
     public double calculateLlcr(Llcr llcr) {
        double LLR = llcr.getLoanLossReserves();
@@ -25,7 +29,15 @@ public class LlcrServiceImpl implements LlcrService {
 
     @Override
     public Llcr addLlcr(Llcr llcr) {
-        return llcrRepository.save(llcr);
+//        return llcrRepository.save(llcr);
+
+    	int count = llcrRepository.countByPeriod(llcr.getPeriod());  
+	      
+	    if (count > 0) {  
+	        throw new DataAlreadyExistsException("Data for period " + llcr.getPeriod() + " already exists. You can only modify it.");  
+	    }  
+	      
+	    return llcrRepository.save(llcr); 
     }
 
     @Override
@@ -35,11 +47,24 @@ public class LlcrServiceImpl implements LlcrService {
 
     @Override
     public void deleteLlcr(int id) {
+    	if (!llcrRepository.existsById(id)) {
+            throw new DataNotFoundException("LLCR with ID " + id + " not found.");
+        }
         llcrRepository.deleteById(id);
     }
 
 	@Override
 	public List<Llcr> allLlcr(int n) {
 		return llcrRepository.getLlcrByMonthPeriod(n);
+	}
+
+	@Override
+	public Llcr editLlcr(int id, Llcr updatedLlcr) {
+		if (!llcrRepository.existsById(id)) {
+            throw new DataNotFoundException("Llcr with ID " + id + " not found.");
+        }
+
+		updatedLlcr.setId(id); // Ensure the ID remains the same
+       return llcrRepository.save(updatedLlcr);
 	}
 }
