+++ ficometer/BankService/src/main/java/com/bank/serviceImpl/RoeServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.bank.exception.DataAlreadyExistsException;
+import com.bank.exception.DataNotFoundException;
 import com.bank.modal.Roe;
 import com.bank.repository.RoeRepo;
 import com.bank.service.RoeService;
@@ -16,7 +18,15 @@ public class RoeServiceImpl implements RoeService{
 	
 	@Override
 	public Roe addRoe(Roe roe) {
-		return roeRepository.save(roe);
+//		return roeRepository.save(roe);
+		
+		int count = roeRepository.countByPeriod(roe.getPeriod());  
+	      
+	    if (count > 0) {  
+	        throw new DataAlreadyExistsException("Data for period " + roe.getPeriod() + " already exists. You can only modify it.");  
+	    }  
+	      
+	    return roeRepository.save(roe); 
 	}
 
 	@Override
@@ -34,8 +44,21 @@ public class RoeServiceImpl implements RoeService{
 	}
 
 	@Override
+    public Roe editRoe(int id, Roe updatedRoe) {
+        if (!roeRepository.existsById(id)) {
+            throw new DataNotFoundException("Roe with ID " + id + " not found.");
+        }
+
+        updatedRoe.setId(id); // Ensure the ID remains the same
+        return roeRepository.save(updatedRoe);
+    }
+	 
+	@Override
 	public void deleteRoe(int id) {
-		roeRepository.deleteById(id);
+		  if (!roeRepository.existsById(id)) {
+	            throw new DataNotFoundException("Roe with ID " + id + " not found.");
+	        }
+	       roeRepository.deleteById(id);
 	}
 
 	@Override
