+++ ficometer/BankService/src/main/java/com/bank/serviceImpl/lcrServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.bank.exception.DataAlreadyExistsException;
+import com.bank.exception.DataNotFoundException;
 import com.bank.modal.Lcr;
 import com.bank.repository.lcrRepo;
 import com.bank.service.lcrService;
@@ -27,7 +29,15 @@ public class lcrServiceImpl implements lcrService{
 
 	@Override
 	public Lcr addLcr(Lcr lcr) {
-		return lcrRepository.save(lcr);
+//		return lcrRepository.save(lcr);
+		
+		int count = lcrRepository.countByPeriod(lcr.getPeriod());  
+	      
+	    if (count > 0) {  
+	        throw new DataAlreadyExistsException("Data for period " + lcr.getPeriod() + " already exists. You can only modify it.");  
+	    }  
+	      
+	    return lcrRepository.save(lcr);  
 	}
 
 	@Override
@@ -37,12 +47,25 @@ public class lcrServiceImpl implements lcrService{
 
 	@Override
 	public void deleteLcr(int id) {
+		if (!lcrRepository.existsById(id)) {
+            throw new DataNotFoundException("LCR with ID " + id + " not found.");
+        }
 		lcrRepository.deleteById(id);
 	}
 
 	@Override
 	public List<Lcr> allLcr(int n) {
 		return lcrRepository.getLcrByMonthPeriod(n);
+	}
+
+	@Override
+	public Lcr editLcr(int id, Lcr updatedLcr) {
+		if (!lcrRepository.existsById(id)) {
+            throw new DataNotFoundException("Llcr with ID " + id + " not found.");
+        }
+
+		updatedLcr.setId(id); // Ensure the ID remains the same
+       return lcrRepository.save(updatedLcr);
 	}
 	
 }
